+++ sdk/communication/azure-communication-jobrouter/src/main/java/com/azure/communication/jobrouter/models/LabelValue.java
@@ -3,10 +3,15 @@
 
 package com.azure.communication.jobrouter.models;
 
+import com.azure.core.util.logging.ClientLogger;
+
 /**
  * Wrapper class for labels. Supports double, String and boolean types.
  */
 public final class LabelValue {
+
+    private static final ClientLogger LOGGER = new ClientLogger(LabelValue.class);
+
 
     /**
      * Value to pass to server.
@@ -17,7 +22,7 @@ public final class LabelValue {
      * Constructor for numerical value.
      * @param numericValue numeric value of label.
      */
-    public LabelValue(double numericValue) {
+    public LabelValue(Double numericValue) {
         this.value = numericValue;
     }
 
@@ -33,7 +38,7 @@ public final class LabelValue {
      * Constructor for boolean value.
      * @param boolValue boolean value of label.
      */
-    public LabelValue(boolean boolValue) {
+    public LabelValue(Boolean boolValue) {
         this.value = boolValue;
     }
 
@@ -43,5 +48,38 @@ public final class LabelValue {
      */
     public Object getValue() {
         return this.value;
+    }
+
+    /**
+     * Returns Double value of object
+     * @return (Double) value.
+     */
+    public Double getValueAsDouble() {
+        if (value.getClass() == Double.class) {
+            return (Double) this.value;
+        }
+        throw LOGGER.logExceptionAsError(new IllegalStateException("value is not of type Double."));
+    }
+
+    /**
+     * Returns String value of object
+     * @return (String) value.
+     */
+    public String getValueAsString() {
+        if (value.getClass() == String.class) {
+            return (String) this.value;
+        }
+        throw LOGGER.logExceptionAsError(new IllegalStateException("value is not of type String."));
+    }
+
+    /**
+     * Returns Boolean value of object
+     * @return (Boolean) value.
+     */
+    public Boolean getValueAsBoolean() {
+        if (value.getClass() == Boolean.class) {
+            return (Boolean) this.value;
+        }
+        throw LOGGER.logExceptionAsError(new IllegalStateException("value is not of type Boolean."));
     }
 }