+++ sdk/cosmos/azure-cosmos-test/src/main/java/com/azure/cosmos/test/faultinjection/FaultInjectionOperationType.java
@@ -35,11 +35,30 @@ public enum FaultInjectionOperationType {
      * Patch item.
      */
     PATCH_ITEM,
-
     /**
      * Batch item.
      */
     BATCH_ITEM
 
     // Add support for metadata request type
+    /**
+     * Read container request.
+     */
+    METADATA_REQUEST_CONTAINER,
+    /**
+     * Read database account request.
+     */
+    METADATA_REQUEST_DATABASE_ACCOUNT,
+    /**
+     * Query query plan request.
+     */
+    METADATA_REQUEST_QUERY_PLAN,
+    /**
+     * Partition key ranges request.
+     */
+    METADATA_REQUEST_PARTITION_KEY_RANGES,
+    /**
+     * Address refresh request.
+     */
+    METADATA_REQUEST_ADDRESS_REFRESH;
 }