+++ sdk/openai/azure-ai-openai/src/samples/java/com/azure/ai/openai/impl/ReadmeSamples.java
@@ -143,7 +143,7 @@ public final class ReadmeSamples {
         ChatCompletions chatCompletions = client.getChatCompletions("{deploymentOrModelId}",
             new ChatCompletionsOptions(chatMessages));
 
-        System.out.printf("Model ID=%s is created at %d.%n", chatCompletions.getId(), chatCompletions.getCreated());
+        System.out.printf("Model ID=%s is created at %d.%n", chatCompletions.getId(), chatCompletions.getCreatedAt());
         for (ChatChoice choice : chatCompletions.getChoices()) {
             ChatMessage message = choice.getMessage();
             System.out.printf("Index: %d, Chat Role: %s.%n", choice.getIndex(), message.getRole());
@@ -165,7 +165,7 @@ public final class ReadmeSamples {
             new ChatCompletionsOptions(chatMessages));
 
         chatCompletionsStream.forEach(chatCompletions -> {
-            System.out.printf("Model ID=%s is created at %d.%n", chatCompletions.getId(), chatCompletions.getCreated());
+            System.out.printf("Model ID=%s is created at %d.%n", chatCompletions.getId(), chatCompletions.getCreatedAt());
             for (ChatChoice choice : chatCompletions.getChoices()) {
                 ChatMessage message = choice.getDelta();
                 if (message != null) {
@@ -193,7 +193,7 @@ public final class ReadmeSamples {
         Embeddings embeddings = client.getEmbeddings("{deploymentOrModelId}", embeddingsOptions);
 
         for (EmbeddingItem item : embeddings.getData()) {
-            System.out.printf("Index: %d.%n", item.getIndex());
+            System.out.printf("Index: %d.%n", item.getPromptIndex());
             for (Double embedding : item.getEmbedding()) {
                 System.out.printf("%f;", embedding);
             }